+++ app/components/Carousel/Carousel.tsx
@@ -2,37 +2,16 @@
 
 import { useState, useEffect, useRef } from 'react';
 import styled, { keyframes, css } from 'styled-components';
-import { CarouselData } from '@/app/data';
+import { CarouselData, CarouselCopyData } from '@/app/data';
 import { CarouselContent } from '@/app/components/Carousel/CarouselContent';
 import { ControlsContainer } from '@/app/components/Carousel/ControlsContainer';
-
-interface ImageContainerProps {
-  active: boolean;
-}
+import { CarouselImages } from '@/app/components/Carousel/CarouselImages';
 
 const CarouselContainer = styled.div`
   position: relative;
   overflow: hidden;
   margin: 15.8rem auto; /* Center the carousel and set top and bottom margins */
   min-height: 530px; /* Set maximum height */
-`;
-
-const fadeIn = keyframes`
-  from {
-    opacity: 0;
-  }
-  to {
-    opacity: 1;
-  }
-`;
-
-const fadeOut = keyframes`
-  from {
-    opacity: 1;
-  }
-  to {
-    opacity: 0;
-  }
 `;
 
 const CarouselInner = styled.div`
@@ -45,71 +24,6 @@ const CarouselInner = styled.div`
   @media (max-width: 768px) {
     width: calc(100% - 20px); /* Adjust width for small devices */
   }
-`;
-
-const CarouselImages = styled.div`
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-  max-width: 518px; /* Set maximum width */
-`;
-
-const ImageContainer = styled.div<ImageContainerProps>`
-  display: ${(props) => (props.active ? 'grid' : 'none')};
-  position: relative;
-  grid-template-columns: 1fr 1fr;
-  grid-template-rows: 1fr 1fr;
-  gap: 15px; /* Add gap between images */
-  height: 530px; /* Set maximum height */
-  max-width: 518px; /* Set maximum width */
-  animation: ${(props) =>
-    props.active
-      ? css`
-          ${fadeIn} 0.5s forwards
-        `
-      : css`
-          ${fadeOut} 0.5s forwards
-        `};
-`;
-
-const ActiveImage = styled.img`
-  grid-column: 1 / 2;
-  grid-row: 1 / 3;
-  justify-self: center;
-  align-self: center;
-  width: 100%;
-  max-width: 28rem;
-  height: 100%;
-  object-fit: scale-down;
-  z-index: 2;
-`;
-
-const InactiveImage = styled.img`
-  grid-column: 2 / 3;
-  grid-row: 1 / 2;
-  justify-self: center;
-  align-self: start;
-  width: 100%;
-  max-width: 28rem;
-  height: 100%;
-  object-fit: scale-down;
-  opacity: 0.3;
-  z-index: 0;
-`;
-
-const InactiveImage2 = styled.img`
-  grid-column: 2 / 3;
-  grid-row: 2 / 3;
-  justify-self: center;
-  align-self: end;
-  position: sticky;
-  bottom: 0;
-  width: 100%;
-  max-width: 28rem;
-  height: 100%;
-  object-fit: scale-down;
-  opacity: 0.3;
-  z-index: 1;
 `;
 
 const Carousel = () => {
@@ -150,11 +64,13 @@ const Carousel = () => {
     setActiveIndex(index);
   };
 
-  const Title = 'Lorem ipsum dolor';
-  const Heading = 'Lorem ipsum dolor sit amet, est mollis evertitur ut';
-  const Copy =
-    'Lorem ipsum dolor sit amet, est mollis evertitur ut, clita utinam quaeque ad sed, an legere concludaturque eum. Duo omnis solet dissentiet te, ea sed quis erat reprehendunt, cetero consetetur philosophia mel in.';
-  const CTACopy = 'Learn More';
+  const CarouselCopy = CarouselCopyData[0];
+  const Title = CarouselCopy.carouselTitle;
+  const Heading = CarouselCopy.carouselHeading;
+  const Copy = CarouselCopy.carouselText;
+  const CTACopy = CarouselCopy.ctaInfo.ctaText;
+  const CTAUrl = CarouselCopy.ctaInfo.ctaUrl;
+  const CTATarget = CarouselCopy.ctaInfo.ctaTarget;
 
   return (
     <CarouselContainer>
@@ -164,17 +80,10 @@ const Carousel = () => {
           heading={Heading}
           text={Copy}
           ctaText={CTACopy}
-          ctaLink='/'
+          ctaLink={CTAUrl}
+          ctaTarget={CTATarget}
         />
-        <CarouselImages>
-          {CarouselData.map((item, index) => (
-            <ImageContainer key={index} active={index === activeIndex}>
-              <ActiveImage src={item.carouselImages.activeImage} />
-              <InactiveImage src={item.carouselImages.inactiveImage1} />
-              <InactiveImage2 src={item.carouselImages.inactiveImage2} />
-            </ImageContainer>
-          ))}
-        </CarouselImages>
+        <CarouselImages data={CarouselData} activeIndex={activeIndex} />
       </CarouselInner>
       <ControlsContainer
         activeIndex={activeIndex}
