+++ app/components/Carousel/CarouselImages/index.js
@@ -1,0 +1,1 @@
+export { default as CarouselImages } from './CarouselImages';
